Move ObjIdSerializer.java to test scope
ObjIdSerializer.java seems to be a utility to help
with testing index serialization logic. It can safely
be moved to test scope.
